cleaning up CSS file
Gave variables to the colors to make them easier to change if anyone wants a different color scheme all they have to do is change the root variables.

Moved text-align center to the body since everything is centered as is, only thing that wouldnt center with that is the input so it has its own.

Deleted the Onsubmit attribute from the HTML file as its didnt actually do anything.

Summary:
Deleted a bunch of un-used properties. Deleted the OnSubmit attributed from the form on infex.html